--- a/parlai/core/torch_agent.py
+++ b/parlai/core/torch_agent.py
@@ -1821,9 +1821,7 @@ class TorchAgent(ABC, Agent):
             if self._is_lr_warming_up():
                 self.warmup_scheduler.step(epoch=self._number_training_updates)
 
-        if self.opt.get('lr_scheduler') == 'invsqrt' and not self._is_lr_warming_up():
-            # training step scheduler
-            self.scheduler.step(self._number_training_updates)
+       
 
         if self.fp16:
             # we've been accumulating grads in fp16 and delaying the fp32 copy update.
@@ -1842,6 +1840,10 @@ class TorchAgent(ABC, Agent):
 
         self.metrics['updates'] += 1
         self.optimizer.step()
+        
+        if self.opt.get('lr_scheduler') == 'invsqrt' and not self._is_lr_warming_up():
+            # training step scheduler
+            self.scheduler.step(self._number_training_updates)
 
     def zero_grad(self):
         """