--- a/parlai/core/torch_agent.py
+++ b/parlai/core/torch_agent.py
@@ -1811,22 +1811,6 @@ class TorchAgent(ABC, Agent):
             if self._number_grad_accum != 0:
                 return
 
-        self.optimizer.step()
-
-        # keep track up number of steps, compute warmup factor
-        self._number_training_updates += 1
-
-        # compute warmup adjustment if needed
-        if self.opt.get('warmup_updates', -1) > 0:
-            if not hasattr(self, 'warmup_scheduler'):
-                raise RuntimeError('Looks like you forgot to call build_lr_scheduler')
-            if self._is_lr_warming_up():
-                self.warmup_scheduler.step(epoch=self._number_training_updates)
-
-        if self.opt.get('lr_scheduler') == 'invsqrt' and not self._is_lr_warming_up():
-            # training step scheduler
-            self.scheduler.step(self._number_training_updates)
-
         if self.fp16:
             # we've been accumulating grads in fp16 and delaying the fp32 copy update.
             # finally time to perform the update.
@@ -1843,6 +1827,21 @@ class TorchAgent(ABC, Agent):
             self.metrics['clip'] += float(grad_norm > self.opt['gradient_clip'])
 
         self.metrics['updates'] += 1
+        self.optimizer.step()
+
+        # keep track up number of steps, compute warmup factor
+        self._number_training_updates += 1
+
+        # compute warmup adjustment if needed
+        if self.opt.get('warmup_updates', -1) > 0:
+            if not hasattr(self, 'warmup_scheduler'):
+                raise RuntimeError('Looks like you forgot to call build_lr_scheduler')
+            if self._is_lr_warming_up():
+                self.warmup_scheduler.step(epoch=self._number_training_updates)
+
+        if self.opt.get('lr_scheduler') == 'invsqrt' and not self._is_lr_warming_up():
+            # training step scheduler
+            self.scheduler.step(self._number_training_updates)
 
     def zero_grad(self):
         """