@stephenroller 2019-10-28
fyi the lines around gradient clipping *must* occur before optimizer.step. Probably you want to move lr_scheduler.step instead.

I think there might be a small issue around the *very* first movement though.