@stephenroller 2019-11-04
Tbh we should have a test but I think this is probably okay. I'm just worried about that very first call to `optimizer.step()`.