@stephenroller 2019-11-06
That the learning rate is exactly what it should be at steps 0, 1, and 10 or something. Particularly with `--lr-scheduler invsqrt` for example.