@stephenroller 2019-11-22
@hadasah when you work on the other LR task, please write some unit tests for the schedulers (isolated) and one integration test.