@klshuster 2019-11-25
Just a note that @domrigoglioso spoke with me offline and indicated he will be writing a test for this