@domrigoglioso 2019-11-25
I just pushed that unit test, it just checks the invsqrt learning rate is right at step 9 and step 16.